+++ src/mesh/test/flame/flame-regression.h
@@ -40,6 +40,7 @@ using namespace ns3;
  * <-----------|----------->   Broadcast frame
  *             |----------->|  Unicast frame
  *           Source                   Destination
+ * (node ID)   2            1            0
  * <-----------|----------->|            |             ARP request
  *             |<-----------|----------->|             ARP request
  *             |            |<-----------|             ARP reply